requirements.txt
tensorflow-cpu==2.7.0
google-cloud-storage==1.42.3
wandb==0.12.5
func-timeout==4.3.5
smart_open[gcs]==5.2.1
jax==0.3.16
jaxlib==0.3.15
optax==0.1.3
dm-haiku==0.0.7
einops==0.3.2